+++ Changelog.md
@@ -191,9 +191,12 @@
   - [#484](../../issues/484) - Add special style/template for MenuItem with set description
   - [#488](../../issues/488) - Display border around content area of RibbonTabControl  
   You can now use `Fluent.Ribbon.Values.RibbonTabControl.Content.BorderThickness` to control the thickness of the border around the content area of `RibbonTabControl`
+  - [#494](../../issues/494) - How to align controls such as ToggleButton, Spinner with text  
+  You can now opt in to align the headers of controls like `ComboBox` or `Spinner` which are placed in the same column of a `RibbonGroupBox` by adding `Grid.IsSharedSizeScope="True"` to a `RibbonGroupBox`.
+  You can opt out of this behavior for single controls in that column by adding `Fluent:RibbonGroupBoxWrapPanel.ExcludeFromSharedSize="True"` to that control.
   - [#495](../../issues/495) - Add option disable handling of KeyTips  
-  You can now disable handling of all KeyTips by setting `IsKeyTipHandlingEnabled` on `Ribbon` to `False`
-  - [#503](../../issues/503) - Add IsDefinitive property to GalleryItem (thanks @noctis0430)
+  You can now disable handling of all KeyTips by setting `IsKeyTipHandlingEnabled` on `Ribbon` to `False`.
+  - [#503](../../issues/503) - Add IsDefinitive property to GalleryItem (thanks @noctis0430)  
 
 ## 5.0.2
   - [#437](../../issues/437) - "Could not load ControlzEx"